requirements.txt
# Core Dependencies
matplotlib>=3.7.0

# PDF Processing
pdfplumber>=0.9.0
camelot-py>=0.11.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0

# Deep Learning and Image Processing
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.23.0

# OCR and Text Processing
pytesseract>=0.3.10
easyocr>=1.7.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
sympy>=1.12.0

# Scientific Computing
scipy>=1.10.0

# Progress and Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2 